Add new sheet for amplifier (post i-v stage)
I'm not sure if I'll stick with this here; my thinking is that it makes
sense to keep the post I-V stage amplifier on the same PCB as the DAC to
minimize noise on the amplifier input. I imagine that PCB interconnects
would cause a fairly significant amount of noise, so I'd like to avoid
that.